+++ product/src/main/java/com/sparta/hotdeal/product/domain/repository/product/ProductRepository.java
@@ -1,6 +1,7 @@
 package com.sparta.hotdeal.product.domain.repository.product;
 
 import com.sparta.hotdeal.product.domain.entity.product.Product;
+import com.sparta.hotdeal.product.infrastructure.repository.ProductRepositoryCustom;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
